# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My-home API",
    "description": "API for my portfolio page",
    "version": "1.0.0"
  },
  "host": "localhost:8080/api",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "accessToken": {
      "type": "apiKey",
      "in": "cookie",
      "name": "accessToken",
      "description": "Need to get access"
    },
    "refreshToken": {
      "type": "apiKey",
      "in": "cookie",
      "name": "refreshToken",
      "description": "Need to refresh tokens"
    }
  },
  "paths": {
...
    "/admin/project": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "create project",
        "description": "Allows the administrator to create a new project",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "New project data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "New Project"
                },
                "description": {
                  "type": "string",
                  "example": "Project desc"
                },
                "img": {
                  "type": "string",
                  "example": "img link"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Result info",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/projects/:id": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "edit project",
        "description": "Allows the administrator to edit a project",
        "parameters": [
          {
            "name": "id",
            "description": "Edited project id",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Edited project data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "delete project",
        "description": "Allows the administrator to delete a project",
        "parameters": [
          {
            "name": "id",
            "description": "Deleted project id",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/projects": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "get projects",
        "description": "Get projects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Response"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/project/:id": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "get project",
        "description": "Get project",
        "parameters": [
          {
            "name": "id",
            "description": "Project id",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Response"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/feedbacks": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "get feedbacks",
        "description": "Get feedbacks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Result info",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Response"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Project": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "New Project"
        },
        "description": {
          "type": "string",
          "example": "Project desc"
        },
        "img": {
          "type": "string",
          "example": "img link"
        }
      }
    },
...
    "Response": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "string"
        },
        "statusCod": {
          "type": "string",
          "example": "number"
        },
        "message": {
          "type": "string",
          "example": "string"
        }
      }
    }
  }
}